app/backend/requirements.txt
fastapi
uvicorn[standard]
Pillow
tensorflow
ultralytics
python-multipart